Fix reboot workaround on svirt backend
failing s390x test with workaround https://openqa.suse.de/tests/610873#step/reboot_gnome/21

+++ tests/x11/reboot_gnome.pm
@@ -40,12 +40,15 @@ sub run() {
 
         send_key "ret";
 
-        wait_still_screen 8, 8;    # wait 8 seconds to make authentication window disappear after successful authentication
-        if (check_screen 'reboot-auth', 2) {
-            record_soft_failure 'bsc#981299';
-            send_key_until_needlematch 'generic-desktop', 'esc',             7, 10;    # close timed out authentication window
-            send_key_until_needlematch 'logoutdialog',    'ctrl-alt-delete', 7, 10;    # reboot
-            assert_and_click 'logoutdialog-reboot-highlighted';
+        # run only on qemu backend, e.g. svirt backend is fast enough to reboot properly
+        if (check_var('BACKEND', 'qemu')) {
+            sleep 10;    # wait 10 seconds to make authentication window disappear after successful authentication
+            if (check_screen 'reboot-auth', 2) {
+                record_soft_failure 'bsc#981299';
+                send_key_until_needlematch 'generic-desktop', 'esc',             7, 10;    # close timed out authentication window
+                send_key_until_needlematch 'logoutdialog',    'ctrl-alt-delete', 7, 10;    # reboot
+                assert_and_click 'logoutdialog-reboot-highlighted';
+            }
         }
     }
     workaround_type_encrypted_passphrase;
